requirements.txt
MouseInfo==0.1.3
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyocr==0.8.5
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
